Update directives to align with PatternFly documentation terminology
This change is not backwards compatible.
The following name changes are made:

pf-utilization-chart to pf-utlization-trend-chart
pf-simple-filter to pf-filter
pf-simple-filter-results to pf-filter-results
pf-notification to pf-inline-notification
pf-simple-sort to pf-sort
pf-data-list to pf-list-view
pf-data-tiles to pf-card-view

CSS selectors have also been changed from pf-<xxx> to <xxx>-pf to be
more inline with PatternFly conventions.

Fixes