PTNFLY-1122 : Fix pfSelect options and ng-model watches
The directive was watching the wrong scope for changes, since it defines
its isolated scope, it will need to watch parent scope.

Also, wrapping selectpicker('refresh') calls in $applyAsync so angular
is quickly aware of changes and shows them.

This fixes 2 problems, not showing changes to options (eg: adding a new
element to options) and not reflecting changes to ng-model (eg: changing
ng-model outside of directive, like with result from REST call).

Fixes 177