Add capability to add app defined action buttons to the toolbar
Currently, there is a limitation in the toolbar. One cannot add, say, a
dropown button to the action buttons. This provides the capability for
an application to add buttons how they want to the toolbar action
buttons area using a named transclusion.

NG-Doc example does not use this feature since its not something we
need to show off and ng docs is not using angular 1.5 which is required
to use this feature.