Add pfVerticalNavigation directive.
Directive for vertical navigation.

This sets up the nav bar header with the collapse button (hamburger)
and the application brand image (or text) as well as the vertical
navigation menu containing the navigation items.

This  directive supports primary, secondary, and tertiary
navigation with options to allow pinning of the secondary and
tertiary navigation menus as well as the option for persistent
secondary menus.